Revert "Feature/update book"
